Move DrawerAdapter to common package
