+++ app/src/main/java/org/gdg/frisbee/android/common/GdgNavDrawerActivity.java
@@ -45,8 +45,7 @@ import org.gdg.frisbee.android.Const;
 import org.gdg.frisbee.android.R;
 import org.gdg.frisbee.android.about.AboutActivity;
 import org.gdg.frisbee.android.activity.SettingsActivity;
-import org.gdg.frisbee.android.adapter.DrawerAdapter;
-import org.gdg.frisbee.android.adapter.DrawerAdapter.DrawerItem;
+import org.gdg.frisbee.android.common.DrawerAdapter.DrawerItem;
 import org.gdg.frisbee.android.api.GapiOkTransport;
 import org.gdg.frisbee.android.app.App;
 import org.gdg.frisbee.android.arrow.ArrowActivity;
